+++ modules/effects/src/index.ts
@@ -52,7 +52,11 @@ export type {
   BloomUniforms
 } from './passes/postprocessing/image-blur-filters/bloom';
 export {bloom} from './passes/postprocessing/image-blur-filters/bloom';
-export {bloomShaderPassPipeline} from './passes/postprocessing/image-blur-filters/bloom-shader-pass-pipeline';
+export type {BloomShaderPassPipelineOptions} from './passes/postprocessing/image-blur-filters/bloom-shader-pass-pipeline';
+export {
+  bloomShaderPassPipeline,
+  createBloomShaderPassPipeline
+} from './passes/postprocessing/image-blur-filters/bloom-shader-pass-pipeline';
 export type {
   DofProps,
   DofUniforms
@@ -129,6 +133,14 @@ export {
   depthAwareBlur,
   depthAwareBlurShaderPassPipeline
 } from './passes/screen-space/depth-aware-blur';
+export type {HDRAutoExposureShaderPassPipelineOptions} from './passes/screen-space/hdr-auto-exposure';
+export {
+  createHDRAutoExposureShaderPassPipeline,
+  hdrAutoExposureAdapt,
+  hdrAutoExposureApply,
+  hdrLuminanceExtract,
+  hdrLuminanceReduce
+} from './passes/screen-space/hdr-auto-exposure';
 export {createMotionBlurShaderPassPipeline} from './passes/screen-space/motion-blur';
 export type {GTAOShaderPassPipelineOptions} from './passes/screen-space/gtao';
 export {